+++ src/metrics/gpu/archs/dummy.c
@@ -15,6 +15,8 @@
 * found in COPYING.BSD and COPYING.EPL files.
 */
 
+//#define SHOW_DEBUGS 1
+#include <common/output/debug.h>
 #include <stdlib.h>
 #include <metrics/gpu/archs/dummy.h>
 
@@ -31,6 +33,7 @@ state_t gpu_dummy_status()
 
 state_t gpu_dummy_init(ctx_t *c)
 {
+	debug("gpu_dummy_init");
 	return EAR_SUCCESS;
 }
 
@@ -41,6 +44,7 @@ state_t gpu_dummy_dispose(ctx_t *c)
 
 state_t gpu_dummy_count(ctx_t *c, uint *_dev_count)
 {
+	debug("count=1");
 	if (_dev_count != NULL) {
 		*_dev_count = 1;
 	}
@@ -49,6 +53,7 @@ state_t gpu_dummy_count(ctx_t *c, uint *_dev_count)
 
 state_t gpu_dummy_read(ctx_t *c, gpu_t *data)
 {
+	debug("read is null");
 	return gpu_dummy_data_null(data);
 }
 
@@ -71,17 +76,19 @@ state_t gpu_dummy_read_copy(ctx_t *c, gpu_t *data2, gpu_t *data1, gpu_t *data_di
 state_t gpu_dummy_data_diff(gpu_t *data2, gpu_t *data1, gpu_t *data_diff)
 {
 	if (data2 == NULL || data1 == NULL || data_diff == NULL) {
+		debug("some pointer in data_diff is null");
 		return_msg(EAR_ERROR, Error.null_data);
 	}
-	return EAR_SUCCESS;
+	return gpu_dummy_data_null(data_diff);
 }
 
 state_t gpu_dummy_data_merge(gpu_t *data_diff, gpu_t *data_merge)
 {
 	if (data_diff == NULL || data_merge == NULL) {
+		debug("some pointer in data_merge is null");
 		return_msg(EAR_ERROR, Error.null_data);
 	}
-	return EAR_SUCCESS;
+	return gpu_dummy_data_null(data_merge);
 }
 
 state_t gpu_dummy_data_init(uint _dev_count)
@@ -92,12 +99,14 @@ state_t gpu_dummy_data_init(uint _dev_count)
 state_t gpu_dummy_data_alloc(gpu_t **data)
 {
 	if (data == NULL) {
+		debug("Error data pointer in alloc is null");
 		return_msg(EAR_ERROR, Error.null_data);
 	}
 	*data = calloc(1, sizeof(gpu_t));
 	if (*data == NULL) {
 		return_msg(EAR_SYSCALL_ERROR, strerror(errno));
 	}
+	debug("Data allocated");
 	return EAR_SUCCESS;
 }
 
@@ -126,7 +135,10 @@ state_t gpu_dummy_data_copy(gpu_t *data_dst, gpu_t *data_src)
 state_t gpu_dummy_data_print(gpu_t *data, int fd)
 {
 	int i;
-
+	if (data == NULL){ 
+		debug("data is null");
+		return_msg(EAR_ERROR, Error.null_data);
+	}
 	for (i = 0; i < 1; ++i)
 	{
 		dprintf(fd, "gpu%u: %0.2lfJ, %0.2lfW, %luMHz, %luMHz, %lu%%, %lu%%, %luº, %luº, %u, %lu\n",
@@ -146,11 +158,17 @@ state_t gpu_dummy_data_tostr(gpu_t *data, char *buffer, int length)
 	int accuml = 0;
 	size_t s;
 	int i;
-
+	if ((data == NULL) || (buffer == NULL)){ 
+		debug("data or buffer are null");
+		return_msg(EAR_ERROR, Error.null_data);
+	}
+	fprintf(stderr,"gpu_dummy_data_tostr en=%lf pow=%lf util=%lu util_mem=%lu\n",data->energy_j,data->power_w,data->util_gpu,data->util_mem);
 	for (i = 0; i < 1 && length > 0; ++i)
 	{
+		fprintf(stderr,"gpu_dummy_data_tostr accessing gpu %u\n",i);
+		
 		s = snprintf(&buffer[accuml], length,
-			"gpu%u: %0.2lfJ, %0.2lfW, %luMHz, %luMHz, %lu%%, %lu%%, %luº, %luº, %u, %lu\n",
+			"gpu%u: %.2lfJ, %0.2lfW, %luMHz, %luMHz, %lu%%, %lu%%, %luº, %luº, %u, %lu\n",
 			i                   ,
 			data[i].energy_j    , data[i].power_w,
 			data[i].freq_gpu_mhz, data[i].freq_mem_mhz,